Fix for OverflowException on x64

+++ src/HidLibrary/HidDevices.cs
@@ -49,7 +49,7 @@ namespace HidLibrary
             var hidClass = HidClassGuid;
             var deviceInfoSet = NativeMethods.SetupDiGetClassDevs(ref hidClass, null, 0, NativeMethods.DIGCF_PRESENT | NativeMethods.DIGCF_DEVICEINTERFACE);
 
-            if (deviceInfoSet.ToInt32() != NativeMethods.INVALID_HANDLE_VALUE)
+            if (deviceInfoSet.ToInt64() != NativeMethods.INVALID_HANDLE_VALUE)
             {
                 var deviceInfoData = CreateDeviceInfoData();
                 var deviceIndex = 0;
